code fixes refactor
