--- src/components/Repositories/Repository.module.scss
@@ -1,162 +1,0 @@
-@import 'variables';
-
-.repository {
-    padding: $spacer / $line-height;
-    border: 1px solid $brand-grey-lighter;
-    background: $brand-white;
-    border-radius: $border-radius;
-    font-size: $font-size-small;
-    text-align: left;
-    width: 100%;
-    margin-bottom: $spacer;
-    display: flex;
-    flex-wrap: wrap;
-
-    p {
-        align-self: flex-start;
-    }
-}
-
-.repositoryName {
-    font-size: $font-size-h4;
-    margin-top: 0;
-    margin-bottom: $spacer / 2;
-    border: 0;
-    padding: 0;
-    width: 100%;
-    position: relative;
-
-    a:first-of-type {
-        color: $brand-black;
-
-        &:hover,
-        &:focus {
-            color: $brand-pink;
-        }
-    }
-
-    svg {
-        width: 1em;
-        height: 1em;
-        display: inline-block;
-        fill: $brand-grey-light;
-        margin-right: $spacer / 8;
-        margin-left: -.2rem;
-        margin-bottom: -.2rem;
-    }
-}
-
-.forkLine {
-    display: inline;
-    margin-left: $spacer / 4;
-    font-size: $font-size-mini;
-    font-family: $font-family-monospace;
-    font-weight: $font-weight-base;
-    color: $brand-grey-light;
-}
-
-.repositoryRelease {
-    font-family: $font-family-base;
-    font-weight: $font-weight-base;
-    font-size: $font-size-small;
-    color: $brand-grey-light;
-    display: inline-block;
-    margin-left: $spacer / 4;
-
-    &:hover,
-    &:focus {
-        transform: none;
-    }
-}
-
-.repositoryMeta {
-    display: flex;
-    align-items: flex-end;
-    align-self: flex-end;
-    width: 100%;
-}
-
-.repositoryLinks {
-    margin: 0;
-    padding: 0;
-    margin-left: -($spacer / $line-height);
-    width: 80%;
-
-    li {
-        display: inline-block;
-        padding: 0;
-        margin-left: $spacer / $line-height;
-
-        &:before {
-            display: none;
-        }
-    }
-
-    a {
-        font-family: $font-family-button;
-        font-weight: $font-weight-bold;
-    }
-}
-
-.repositorynumbers {
-    font-size: $font-size-small;
-    margin-left: -($spacer / 4);
-    flex: 1;
-    text-align: right;
-    width: 20%;
-
-    a {
-        color: $brand-grey-light;
-        margin-left: $spacer / 4;
-        display: inline-block;
-
-        &:hover,
-        &:focus {
-            color: $brand-pink;
-            transform: none;
-
-            svg {
-                fill: $brand-pink;
-            }
-        }
-
-        span {
-            margin-left: -.15rem;
-        }
-    }
-
-    svg {
-        display: inline-block;
-        fill: $brand-grey-light;
-        margin-bottom: -.1rem;
-    }
-}
-
-.repositoryReadme {
-    margin-top: $spacer / $line-height;
-    padding: $spacer / 2;
-    background: darken($brand-white, 2%);
-
-    h1,
-    h2 {
-        font-size: $font-size-large;
-    }
-
-    h3,
-    h4 {
-        font-size: $font-size-base;
-    }
-
-    img {
-        display: inline-block;
-    }
-}
-
-.repositoryReadmeTitle {
-    font-size: $font-size-small !important; // stylelint-disable-line
-    font-family: $font-family-monospace;
-    font-weight: $font-weight-base;
-    margin: 0;
-    color: $brand-grey-light;
-    margin-bottom: $spacer / 2;
-}
