+++ src/components/Repository/Readme.module.scss
@@ -1,0 +1,30 @@
+@import 'variables';
+
+.readme {
+    margin-top: $spacer / $line-height;
+    padding: $spacer / 2;
+    background: darken($brand-white, 2%);
+
+    h1,
+    h2 {
+        font-size: $font-size-large;
+    }
+
+    h3,
+    h4 {
+        font-size: $font-size-base;
+    }
+
+    img {
+        display: inline-block;
+    }
+}
+
+.title {
+    font-size: $font-size-small !important; // stylelint-disable-line
+    font-family: $font-family-monospace;
+    font-weight: $font-weight-base;
+    margin: 0;
+    color: $brand-grey-light;
+    margin-bottom: $spacer / 2;
+}
