+++ src/components/Repository/Title.module.scss
@@ -1,0 +1,53 @@
+@import 'variables';
+
+.title {
+    font-size: $font-size-h4;
+    margin-top: 0;
+    margin-bottom: $spacer / 2;
+    border: 0;
+    padding: 0;
+    width: 100%;
+    position: relative;
+
+    a:first-of-type {
+        color: $brand-black;
+
+        &:hover,
+        &:focus {
+            color: $brand-pink;
+        }
+    }
+
+    svg {
+        width: 1em;
+        height: 1em;
+        display: inline-block;
+        fill: $brand-grey-light;
+        margin-right: $spacer / 8;
+        margin-left: -.2rem;
+        margin-bottom: -.2rem;
+    }
+}
+
+.repositoryRelease {
+    font-family: $font-family-base;
+    font-weight: $font-weight-base;
+    font-size: $font-size-small;
+    color: $brand-grey-light;
+    display: inline-block;
+    margin-left: $spacer / 4;
+
+    &:hover,
+    &:focus {
+        transform: none;
+    }
+}
+
+.forkLine {
+    display: inline;
+    margin-left: $spacer / 4;
+    font-size: $font-size-mini;
+    font-family: $font-family-monospace;
+    font-weight: $font-weight-base;
+    color: $brand-grey-light;
+}
