+++ src/components/Repository/index.jsx
@@ -1,12 +1,11 @@
-import React, { PureComponent } from 'react'
+import React from 'react'
 import PropTypes from 'prop-types'
 import { StaticQuery, graphql } from 'gatsby'
-import remark from 'remark'
-import remarkReact from 'remark-react'
-import axios from 'axios'
-import { ReactComponent as Star } from '../../images/star.svg'
-import { ReactComponent as Forks } from '../../images/forks.svg'
-import styles from './Repository.module.scss'
+import Numbers from './Numbers'
+import Readme from './Readme'
+import styles from './index.module.scss'
+import Title from './Title'
+import Links from './Links'
 
 const queryGithub = graphql`
     query GitHubReposInfo {
@@ -65,155 +64,6 @@ const queryGithub = graphql`
         }
     }
 `
-
-const Title = ({ name, isFork, parent, releases, url }) => (
-    <h1 className={styles.repositoryName}>
-        <a href={url}>
-            {isFork && <Forks />}
-            {name}
-            {isFork && (
-                <span className={styles.forkLine}>{parent.nameWithOwner}</span>
-            )}
-        </a>
-        {releases.edges[0] && (
-            <a
-                href={`${url}/releases`}
-                className={styles.repositoryRelease}
-                title="Latest release"
-            >
-                {releases.edges[0].node.tag.name}
-            </a>
-        )}
-    </h1>
-)
-
-Title.propTypes = {
-    name: PropTypes.string.isRequired,
-    isFork: PropTypes.bool,
-    parent: PropTypes.shape({
-        nameWithOwner: PropTypes.string
-    }),
-    releases: PropTypes.object.isRequired,
-    url: PropTypes.string.isRequired
-}
-
-const Links = ({ links, url }) => (
-    <ul className={styles.repositoryLinks}>
-        <li>
-            <a href={url}>GitHub</a>
-        </li>
-        {links &&
-            links.map(link => (
-                <li key={link.url}>
-                    <a href={link.url}>{link.name}</a>
-                </li>
-            ))}
-    </ul>
-)
-
-Links.propTypes = {
-    links: PropTypes.array,
-    url: PropTypes.string.isRequired
-}
-
-class Numbers extends PureComponent {
-    static propTypes = {
-        stargazers: PropTypes.object.isRequired,
-        forkCount: PropTypes.number.isRequired,
-        url: PropTypes.string.isRequired,
-        name: PropTypes.string.isRequired
-    }
-
-    state = {
-        forks: this.props.forkCount,
-        stars: this.props.stargazers.totalCount
-    }
-
-    url = 'https://oceanprotocol-github.now.sh'
-
-    signal = axios.CancelToken.source()
-
-    componentDidMount() {
-        this.fetchNumbers()
-    }
-
-    componentWillUnmount() {
-        this.signal.cancel()
-    }
-
-    fetchNumbers = async () => {
-        try {
-            const response = await axios({
-                method: 'get',
-                url: this.url,
-                timeout: 10000, // 10 sec.
-                headers: {
-                    'Content-Type': 'application/json'
-                }
-            })
-
-            const repo = response.data
-                .map(item => {
-                    if (item.name === this.props.name) {
-                        return item
-                    }
-                })
-                .filter(n => n)
-
-            const { forks, stars } = repo
-
-            // update state only when numbers have changed
-            if (forks && forks !== this.props.forkCount) {
-                this.setState({ forks })
-            }
-
-            if (stars && stars !== this.props.stargazers.totalCount) {
-                this.setState({ stars })
-            }
-        } catch (error) {
-            if (axios.isCancel(error)) {
-                return null
-            } else {
-                console.log(error.message) // eslint-disable-line no-console
-            }
-        }
-    }
-
-    render() {
-        const { url } = this.props
-        const { forks, stars } = this.state
-
-        return (
-            <aside className={styles.repositorynumbers}>
-                <a href={`${url}/stargazers`} title="Show Stargazers">
-                    <Star /> <span>{stars}</span>
-                </a>
-                {forks > 0 && (
-                    <a href={`${url}/network`} title="Show Forks">
-                        <Forks /> <span>{forks}</span>
-                    </a>
-                )}
-            </aside>
-        )
-    }
-}
-
-const Readme = ({ object }) => {
-    const readmeHtml =
-        object &&
-        remark()
-            .use(remarkReact)
-            .processSync(object.text).contents
-
-    return (
-        object && (
-            <aside className={styles.repositoryReadme}>
-                <h3 className={styles.repositoryReadmeTitle}>README.md</h3>
-                {readmeHtml}
-            </aside>
-        )
-    )
-}
 
 const Repository = ({ name, links, readme }) => (
     <StaticQuery
